create custom hook useplayer
