+++ api/fetchTracks.js
@@ -1,14 +1,9 @@
 import routes from '../config/routes';
+import fetchApi from './fetchApi';
 
 export default async function () {
   try {
-    const response = await fetch(`${process.env['TIMAPULAN_API_URL']}${routes.GET_TRACKS}`, { 
-      headers: 
-        {
-          'Authorization': process.env['TIMAPULAN_ID_KEY'],
-          'Content-Type': 'application/json'
-        } 
-    });
+    const response = await fetchApi(routes.GET_TRACKS);
     const data = await response.json();
     return data;
   } catch (error) {
